Backend initial setup complete

+++ comments/views.py
@@ -2,14 +2,17 @@ from .serializers.common import CommentSerializer
 from .models import Comment
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
 from rest_framework import status
+from rest_framework.permissions import IsAuthenticated
 
 
 class CommentListView(APIView):
+    permission_classes = (IsAuthenticated, )
     def post(self, request):
-        new_comment = CommentSerializer(data=request.data)
+        request.data['owner'] = request.user.id
         try:
+          new_comment = CommentSerializer(data=request.data)
           if new_comment.is_valid():
             new_comment.save()
             return Response(new_comment.data, status.HTTP_201_CREATED)
@@ -19,9 +22,14 @@ class CommentListView(APIView):
 
 
 class CommentDetailView(APIView):
+    permission_classes = (IsAuthenticated, )
     def delete(self, request, pk):
         try:
           comment = Comment.objects.get(pk=pk)
+
+          if comment.owner != request.user:
+            raise PermissionDenied('Unauthorised')
+
           comment.delete()
           return Response(status=status.HTTP_204_NO_CONTENT)
         except Comment.DoesNotExist as e:
@@ -31,6 +39,10 @@ class CommentDetailView(APIView):
     def put(self, request, pk):
         try:
           comment = Comment.objects.get(pk=pk)
+
+          if comment.owner != request.user:
+            raise PermissionDenied('Unauthorised')
+
           comment = CommentSerializer(comment, request.data, partial=True)
           if comment.is_valid():
             comment.save()
